requirements.txt
google-adk
litellm
nest_asyncio
rich
python-dotenv